Use forEach
