+++ packages/view-types/spatial-three/src/SpatialThreeSubscriber.js
@@ -787,7 +787,6 @@ export function SpatialThreeSubscriber(props) {
       isReady={isReady}
     >
       <SpatialThree
-        ref={deckRef}
         uuid={uuid}
         width={width}
         height={height}
